examples/chirpstack-application-server/integrations/gcppubsub/python/requirements.txt
chirpstack-api==3.7.2
google-cloud-pubsub==1.7.0
